chore: more lint
